refactor: move `scopeHasLocalReference` util function to deduplicate internal code
