+++ src/rules/utils.ts
@@ -672,165 +672,4 @@ export const getTestCallExpressionsFromDeclaredVariables = (
   );
 };
 
-interface ImportDetails {
-  source: string;
-  local: string;
-  imported: string;
-}
-
-const describeImportDefAsImport = (
-  def: TSESLint.Scope.Definitions.ImportBindingDefinition,
-): ImportDetails | null => {
-  if (def.parent.type === AST_NODE_TYPES.TSImportEqualsDeclaration) {
-    return null;
-  }
-
-  if (def.node.type !== AST_NODE_TYPES.ImportSpecifier) {
-    return null;
-  }
-
-  // we only care about value imports
-  if (def.parent.importKind === 'type') {
-    return null;
-  }
-
-  return {
-    source: def.parent.source.value,
-    imported: def.node.imported.name,
-    local: def.node.local.name,
-  };
-};
-
-/**
- * Attempts to find the node that represents the import source for the
- * given expression node, if it looks like it's an import.
- *
- * If no such node can be found (e.g. because the expression doesn't look
- * like an import), then `null` is returned instead.
- */
-const findImportSourceNode = (
-  node: TSESTree.Expression,
-): TSESTree.Node | null => {
-  if (node.type === AST_NODE_TYPES.AwaitExpression) {
-    if (node.argument.type === AST_NODE_TYPES.ImportExpression) {
-      return (node.argument as TSESTree.ImportExpression).source;
-    }
-
-    return null;
-  }
-
-  if (
-    node.type === AST_NODE_TYPES.CallExpression &&
-    isIdentifier(node.callee, 'require')
-  ) {
-    return node.arguments[0] ?? null;
-  }
-
-  return null;
-};
-
-const describeVariableDefAsImport = (
-  def: TSESLint.Scope.Definitions.VariableDefinition,
-): ImportDetails | null => {
-  // make sure that we've actually being assigned a value
-  if (!def.node.init) {
-    return null;
-  }
-
-  const sourceNode = findImportSourceNode(def.node.init);
-
-  if (!sourceNode || !isStringNode(sourceNode)) {
-    return null;
-  }
-
-  if (def.name.parent?.type !== AST_NODE_TYPES.Property) {
-    return null;
-  }
-
-  if (!isSupportedAccessor(def.name.parent.key)) {
-    return null;
-  }
-
-  return {
-    source: getStringValue(sourceNode),
-    imported: getAccessorValue(def.name.parent.key),
-    local: def.name.name,
-  };
-};
-
-/**
- * Attempts to describe a definition as an import if possible.
- *
- * If the definition is an import binding, it's described as you'd expect.
- * If the definition is a variable, then we try and determine if it's either
- * a dynamic `import()` or otherwise a call to `require()`.
- *
- * If it's neither of these, `null` is returned to indicate that the definition
- * is not describable as an import of any kind.
- */
-const describePossibleImportDef = (def: TSESLint.Scope.Definition) => {
-  if (def.type === 'Variable') {
-    return describeVariableDefAsImport(def);
-  }
-
-  if (def.type === 'ImportBinding') {
-    return describeImportDefAsImport(def);
-  }
-
-  return null;
-};
-
-const collectReferences = (scope: TSESLint.Scope.Scope) => {
-  const locals = new Set();
-  const imports = new Map<string, ImportDetails>();
-  const unresolved = new Set();
-
-  let currentScope: TSESLint.Scope.Scope | null = scope;
-
-  while (currentScope !== null) {
-    for (const ref of currentScope.variables) {
-      if (ref.defs.length === 0) {
-        continue;
-      }
-
-      const def = ref.defs[ref.defs.length - 1];
-
-      const importDetails = describePossibleImportDef(def);
-
-      if (importDetails) {
-        imports.set(importDetails.local, importDetails);
-
-        continue;
-      }
-
-      locals.add(ref.name);
-    }
-
-    for (const ref of currentScope.through) {
-      unresolved.add(ref.identifier.name);
-    }
-
-    currentScope = currentScope.upper;
-  }
-
-  return { locals, imports, unresolved };
-};
-
-export const scopeHasLocalReference = (
-  scope: TSESLint.Scope.Scope,
-  referenceName: string,
-) => {
-  const references = collectReferences(scope);
-
-  return (
-    // referenceName was found as a local variable or function declaration.
-    references.locals.has(referenceName) ||
-    // referenceName was found as an imported identifier
-    references.imports.has(referenceName) ||
-    // referenceName was not found as an unresolved reference,
-    // meaning it is likely not an implicit global reference.
-    !references.unresolved.has(referenceName)
-  );
-};
-
 export * from './utils/parseJestFnCall';
