+++ src/rules/utils/parseJestFnCall.ts
@@ -392,3 +392,20 @@ const resolveToJestFn = (
     type: 'global',
   };
 };
+
+export const scopeHasLocalReference = (
+  scope: TSESLint.Scope.Scope,
+  referenceName: string,
+) => {
+  const references = collectReferences(scope);
+
+  return (
+    // referenceName was found as a local variable or function declaration.
+    references.locals.has(referenceName) ||
+    // referenceName was found as an imported identifier
+    references.imports.has(referenceName) ||
+    // referenceName was not found as an unresolved reference,
+    // meaning it is likely not an implicit global reference.
+    !references.unresolved.has(referenceName)
+  );
+};
